+++ src/libopenrave-core/generictrajectory.cpp
@@ -140,7 +140,8 @@ public:
         if( data.size() == 0 ) {
             return;
         }
-        BOOST_ASSERT((data.size()%_spec.GetDOF()) == 0);
+        BOOST_ASSERT(_spec.GetDOF()>0);
+        OPENRAVE_ASSERT_FORMAT((data.size()%_spec.GetDOF()) == 0, "%d does not divide dof %d", data.size()%_spec.GetDOF(), ORE_InvalidArguments);
         if( bOverwrite && index*_spec.GetDOF() < _vtrajdata.size() ) {
             size_t copysize = min(data.size(),_vtrajdata.size()-index*_spec.GetDOF());
             std::copy(data.begin(),data.begin()+copysize,_vtrajdata.begin()+index*_spec.GetDOF());
@@ -160,7 +161,8 @@ public:
         if( data.size() == 0 ) {
             return;
         }
-        BOOST_ASSERT(spec.GetDOF()>0 && (data.size()%spec.GetDOF()) == 0);
+        BOOST_ASSERT(_spec.GetDOF()>0);
+        OPENRAVE_ASSERT_FORMAT((data.size()%_spec.GetDOF()) == 0, "%d does not divide dof %d", data.size()%_spec.GetDOF(), ORE_InvalidArguments);
         if( _spec == spec ) {
             Insert(index,data,bOverwrite);
         }